+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fc-toolkit",
-  "version": "1.0.1",
+  "version": "1.1.0-alpha.1",
   "description": "fc-toolkit",
   "license": "MIT",
   "repository": "",
@@ -35,7 +35,7 @@
   "dependencies": {
     "@alicloud/fc2": "^2.0.4",
     "async-retry": "^1.2.3",
-    "awos-js": "^1.0.3",
+    "awos-js": "^1.0.4",
     "filesize": "^6.0.0",
     "lodash.isnil": "^4.0.0",
     "lodash.isplainobject": "^4.0.6",
